+++ .idea/misc.xml
@@ -19,10 +19,15 @@
         <entry key="..\:/Users/Emmanuel Ikechukwu/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/fragment_splash_screen.xml" value="0.17391304347826086" />
         <entry key="..\:/Users/Emmanuel Ikechukwu/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/screen_splash.xml" value="0.22327898550724637" />
         <entry key="..\:/Users/Emmanuel Ikechukwu/AndroidStudioProjects/Emmy_iBook/app/src/main/res/menu/menu.xml" value="0.24947916666666667" />
+        <entry key="..\:/Users/Peace/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/activity_home_screen.xml" value="0.10989583333333333" />
+        <entry key="..\:/Users/Peace/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/activity_main.xml" value="0.10989583333333333" />
+        <entry key="..\:/Users/Peace/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/fragment_donate.xml" value="0.25" />
+        <entry key="..\:/Users/Peace/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/fragment_signin_page.xml" value="0.22447916666666667" />
+        <entry key="..\:/Users/Peace/AndroidStudioProjects/Emmy_iBook/app/src/main/res/layout/style.xml" value="0.22447916666666667" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
