+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
             android:name=".HomeScreen"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true"
             android:theme="@style/SplashScreenTheme">
             <intent-filter>
