+++ app/src/main/java/com/example/emmy_ibook/activity/HomeScreen.kt
@@ -1,0 +1,17 @@
+package com.example.emmy_ibook.activity
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import com.example.emmy_ibook.R
+import com.example.emmy_ibook.databinding.ActivityMainBinding
+
+class HomeScreen : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_home_screen)
+    }
+}
+
