+++ app/src/main/java/com/example/emmy_ibook/fragments/DonateFragment.kt
@@ -1,0 +1,19 @@
+package com.example.emmy_ibook.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.emmy_ibook.R
+
+class DonateFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_donate, container, false)
+    }
+}
