+++ app/src/main/java/com/example/emmy_ibook/fragments/ProceedPage.kt
@@ -29,7 +29,8 @@ class ProceedPage : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         binding.proceedBtn.setOnClickListener {
-            navCon.navigate(R.id.action_proceedPage_to_signinPage)
+            navCon.navigate(R.id.action_proceedPage_to_donateFragment)
         }
+        binding.proceedBtn.setOnClickListener { navCon.navigate(R.id.action_proceedPage_to_donateFragment) }
     }
 }
