+++ app/src/main/res/values/strings.xml
@@ -27,4 +27,20 @@
     <string name="create_a_username">Create a username</string>
     <string name="provide_your_last_name">Provide your last name</string>
     <string name="provide_your_first_name">Provide your first name</string>
+    <string name="back_arrow">back arrow</string>
+    <string name="donate_a_book">Donate a Book</string>
+    <string name="book_name">Book Name</string>
+    <string name="the_jingle_jangle">The Jingle Jangle</string>
+    <string name="author_name">Author Name</string>
+    <string name="year_published">Year Published</string>
+    <string name="_2017">2017</string>
+    <string name="book_category">Book Category</string>
+    <string name="fiction">Fiction</string>
+    <string name="pickup_location">Pickup Location</string>
+    <string name="card_image">card image</string>
+    <string name="pickup_date">Pickup Date</string>
+    <string name="_10_10_10">10/10/10</string>
+    <string name="pickup_time">Pickup Time</string>
+    <string name="_10_00am">10:00am</string>
+    <string name="complete">Complete</string>
 </resources>
